seperated budget circle componenet
